requirements.txt
# Core AutoGen packages (new structure)
autogen-agentchat>=0.4.0
autogen-ext[openai]>=0.4.0

# Alternative: use pyautogen 0.2.x for stability
# pyautogen~=0.2.0

# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.28.0
pandas>=2.0.0
aiohttp>=3.8.0
colorama>=0.4.6
rich>=13.0.0
pydantic>=2.0.0
tiktoken>=0.5.0

# Research Tool Dependencies
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Optional: AutoGen Studio for GUI
# autogenstudio>=0.1.0

# Web API Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0